빌드 파일 수정 게시글 목록 수정 댓글 개수 표시 스프링 부트 적용 + 패스워드 암호화 적용(해시 함수)
회원가입 폼 구현 중,,

+++ src/main/java/sijang/sijanggaza/repository/UserRepository.java
@@ -3,6 +3,6 @@ package sijang.sijanggaza.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sijang.sijanggaza.domain.SiteUser;
 
-public interface UserRepository extends JpaRepository<SiteUser, Integer> {
+public interface UserRepository extends JpaRepository<SiteUser, Long> {
 
 }
